py_to_linux/mango-pi/requirements.txt
# ===============================
# Core numerical (MATCH TRAINING)
# ===============================
numpy==1.23.5
scipy==1.15.3
scikit-learn==1.3.2
joblib==1.5.2

# ===============================
# Computer Vision
# ===============================
opencv-python==4.8.1.78
Pillow==10.1.0

# ===============================
# Web App
# ===============================
Flask==3.0.0
Werkzeug==3.1.5
Jinja2==3.1.6
click==8.3.1
itsdangerous==2.2.0

# ===============================
# Utilities
# ===============================
tqdm==4.67.1
typing_extensions==4.15.0